% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_course.R
\name{install_workshop}
\alias{install_workshop}
\title{Install workshop materials}
\usage{
install_workshop(path = ".")
}
\arguments{
\item{path}{The path on your computer where you would like the workshop
installed}
}
\description{
`install_workshop()` will install the materials for the workshop "Designing
ggplots" on your computer. Then, it will open a new RStudio Project
containing the files you'll need.
}
